# Docs/openapi_real.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alling B2B API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/formatos/{id}/kits": {
      "post": {
        "tags": [
          "Formato Único"
        ],
        "summary": "Agregar Kit Al Formato",
        "description": "Agrega todos los componentes de un kit al Formato Único.",
        "operationId": "agregar_kit_al_formato_formatos__id__kits_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarKitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormatoResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kits/{id}/": {
      "get": {
        "tags": [
          "Kits"
        ],
        "summary": "Obtener Detalle Kit",
        "description": "Obtiene el detalle de un Kit pre-armado calculando dinámicamente su precio y stock.",
        "operationId": "obtener_detalle_kit_kits__id___get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/favoritos/{producto_id}": {
      "post": {
        "tags": [
          "Favoritos"
        ],
        "summary": "Add Favorite",
        "operationId": "add_favorite_favoritos__producto_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "producto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Producto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/favoritos/": {
      "get": {
        "tags": [
          "Favoritos"
        ],
        "summary": "List Favorites",
        "operationId": "list_favorites_favoritos__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/admin/kits": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Listar Kits Admin",
        "description": "RF-ADM-009: Lista todos los kits (MVP: in-memory).",
        "operationId": "listar_kits_admin_admin_kits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Crear Kit",
        "description": "RF-ADM-009: Crea un nuevo kit de productos.\nBTN-ADM-009: mínimo 2 componentes.",
        "operationId": "crear_kit_admin_kits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KitCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/cotizaciones": {
      "get": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Listar Cotizaciones",
        "description": "RF-COT-001: Lista FormatosUnicos en estados de cotización.\nOPS-COT-001: visibilidad compartida — todos los SELLERs ven todas las cotizaciones.",
        "operationId": "listar_cotizaciones_cotizaciones_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por estado: COTIZACION | EXPIRADA | PEDIDO | CONFIRMADO",
              "title": "Estado"
            },
            "description": "Filtro por estado: COTIZACION | EXPIRADA | PEDIDO | CONFIRMADO"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CotizacionListSchema"
                  },
                  "title": "Response Listar Cotizaciones Cotizaciones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cotizaciones/{fu_id}": {
      "get": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Detalle Cotizacion",
        "description": "RF-COT-002: Devuelve el detalle completo de una cotización.\nOPS-COT-002: sin restricción de asignación (a diferencia de consultas).",
        "operationId": "detalle_cotizacion_cotizaciones__fu_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cotizaciones/{fu_id}/pdf": {
      "get": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Descargar Pdf",
        "description": "RF-COT-003: Redirige al PDF de la cotización.\nBTN-COT-001: solo si pdf_url no es None.\nHTTP 404 si el PDF no existe.",
        "operationId": "descargar_pdf_cotizaciones__fu_id__pdf_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgregarKitRequest": {
        "properties": {
          "kit_id": {
            "type": "string",
            "format": "uuid",
            "title": "Kit Id"
          },
          "cantidad": {
            "type": "integer",
            "title": "Cantidad",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "kit_id"
        ],
        "title": "AgregarKitRequest"
      },
...
      "CotizacionDetailSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "title": "CotizacionDetailSchema"
      },
      "CotizacionListSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "has_pdf": {
            "type": "boolean",
            "title": "Has Pdf"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "has_pdf"
        ],
        "title": "CotizacionListSchema"
      },
...
      "FavoriteResponse": {
        "properties": {
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Ids"
          }
        },
        "type": "object",
        "required": [
          "product_ids"
        ],
        "title": "FavoriteResponse"
      },
      "FormatoResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FormatoUnicoItemSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state",
          "items",
          "subtotal"
        ],
        "title": "FormatoResponseSchema"
      },
      "FormatoUnicoItemSchema": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "stock_disponible": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Disponible"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity",
          "unit_price",
          "subtotal"
        ],
        "title": "FormatoUnicoItemSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KitComponentSchema": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "quantity"
        ],
        "title": "KitComponentSchema"
      },
      "KitCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "component_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 2,
            "title": "Component Ids",
            "description": "Minimum 2 components (BTN-ADM-009)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "component_ids"
        ],
        "title": "KitCreateSchema"
      },
      "KitResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/KitComponentSchema"
            },
            "type": "array",
            "title": "Components"
          },
          "precio_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Precio Total"
          },
          "stock_disponible": {
            "type": "integer",
            "title": "Stock Disponible"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "components",
          "precio_total",
          "stock_disponible"
        ],
        "title": "KitResponseSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}